Software that was created for a client based on their interviews
Software that was created for a client based on their interviews, and
using the skill of communicating with that client in a simple and
non-technical manner. This information then had to be delivered
successfully to a group of technical developers in a more complex and
technical manner. The software allows the user to login, and based on
their login credentials give them certain permissions to be able to
access/edit files based on where they stand in the hierarchy of the
design. This program was achieved with a team of 4 using the agile
method. (Java, MYSQL)